requirements.txt
llama-index==0.9.11
langchain==0.0.312
transformers==4.35.0
chromadb==0.4.18
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
streamlit==1.29.0
pandas==2.1.1
numpy==1.25.2
matplotlib==3.8.0
seaborn==0.13.0
python-dotenv==1.0.0
python-multipart==0.0.6
torch==2.1.0
sentence-transformers==2.2.2
accelerate==0.20.3
protobuf==3.20.0
python-multipart==0.0.6
